数据源连接： 接口修改 数据查询： 接口修改

+++ fisk-chartvisual/fisk-chartvisual-pojo/src/main/java/com/fisk/chartvisual/vo/DataServiceVO.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class DataServiceVO {
     public String name;
-    public Integer value;
+    public Long value;
 }
